pyproject.toml
[project]
name = "cedar-mcp"
version = "0.1.0"
description = "A CEDAR MCP server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "pydantic>=2.0.0",
    "PyYAML>=6.0.0",
]

[project.scripts]
cedar-mcp = "cedar_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cedar_mcp"]